+++ Samples~/Scripts/SaintsEditor/Testing/RemoveMe.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
+{
+    [InfoBox("You can remove component by click \"Destroy And Save\"")]
+    public class RemoveMe : SaintsMonoBehaviour
+    {
+    }
+}